refactor: Implement deeper structural improvements (Part 3)
This commit implements a third round of structural improvements:

- **Code Organization (packages/cli/src/config/extension-source-resolver.ts):**
  Defined an `ExtensionSourceResolver` interface to introduce a more explicit strategy pattern for handling different extension source types.

- **Overall Repository Layout (scripts/generate-git-commit-info.js & scripts/utils/git-utils.js):**
  Extracted the `getGitCommitHash` function into a new utility module `scripts/utils/git-utils.js` and updated `generate-git-commit-info.js` to use it, improving script modularity.

- **Dependency Management (packages/cli/package.json & third_party/get-ripgrep/package.json):**
  Continued standardizing dependency versions to exact versions in `packages/cli/package.json` (for numerous `dependencies` entries) and in `third_party/get-ripgrep/package.json` (for all `dependencies` and `devDependencies`).

- **Build Process (.github/workflows/ci.yml):**
  Added an `if` condition to the `lint` job in `ci.yml` to ensure it runs only when the workflow is triggered by relevant events, contributing to more granular CI control.

- **React Guidelines (packages/cli/src/ui/components/InputPrompt.tsx & InputContentDisplay.tsx):**
  Extracted the complex rendering logic for input lines and ghost text from `InputPrompt.tsx` into a new `InputContentDisplay.tsx` component, significantly reducing the complexity of `InputPrompt.tsx`.
